Fix type checking for GitHub

+++ src/stockbeta/core.py
@@ -14,13 +14,13 @@ def calculate_beta(stock_returns: pd.Series, factor_returns: pd.Series) -> float
     """Calculate beta with respect to any factor."""
     cov = np.cov(stock_returns, factor_returns)[0, 1]
     var = np.var(factor_returns)
-    return cov / var
+    return float(cov / var)
 
 
 def calculate_sharpe_ratio(returns: pd.Series, risk_free_rate: pd.Series) -> float:
     """Calculate the Sharpe ratio for a series of returns."""
     excess_returns = returns - risk_free_rate
-    return np.mean(excess_returns) / np.std(excess_returns)
+    return float(np.mean(excess_returns) / np.std(excess_returns))
 
 
 def calculate_factor_exposures(stock_returns: pd.Series, factors: pd.DataFrame) -> dict:
